+++ hash_dictionary/hashtable.h
@@ -22,26 +22,26 @@
 typedef struct htabl hashtable;
 struct htabl
 {
-    size_t (*hash)(const void*, const hashtable *);
+    int (*hash)(const void*, const hashtable *);
     int (*compare)(const void*, const void*);
     void (*freeObject)(void*);
     void* *entries;
-    size_t size;
+    int size;
 };
 
 hashtable ht_create_empty();
-size_t ht_init(hashtable *ht, size_t size, size_t (*hash)(const void*, const hashtable *), int (*compare)(const void*, const void*), void (*freeObject)(void*));
+int ht_init(hashtable *ht, int size, int (*hash)(const void*, const hashtable *), int (*compare)(const void*, const void*), void (*freeObject)(void*));
 bool ht_destroy(hashtable *ht);
 bool ht_free(hashtable *ht);
-size_t ht_trim(hashtable *ht);
+int ht_trim(hashtable *ht);
 sdarray ht_to_array(hashtable *ht);
-hashtable ht_from_array(sdarray* a, size_t (*hash)(const void*, const hashtable *), int (*compare)(const void*, const void*), void (*freeObject)(void*));
+hashtable ht_from_array(sdarray* a, int (*hash)(const void*, const hashtable *), int (*compare)(const void*, const void*), void (*freeObject)(void*));
 
 int ht_insert(hashtable *ht, void* element);
 void* ht_delete(hashtable *ht, const void* element);
 void* ht_lookup(const hashtable *ht, const void* element);
 
-size_t ht_size(const hashtable *ht);
-size_t ht_count(const hashtable *ht);
+int ht_size(const hashtable *ht);
+int ht_count(const hashtable *ht);
 
 #endif
