+++ utils/file.h
@@ -17,7 +17,7 @@ write/read files
  * @param contents a pointer to the buffer where the contents of the file will be written to
  * @return bytes read
  */
-size_t read_file(const char *filepath, void **contents);
+int read_file(const char *filepath, void **contents);
 
 /**
  * write data to file (binary)
@@ -27,7 +27,7 @@ size_t read_file(const char *filepath, void **contents);
  * @param size the size of the contents in bytes
  * @return bytes written
  */
-size_t write_file(const char *filepath, void *contents, size_t size);
+int write_file(const char *filepath, void *contents, uint size);
 
 /**
  * @brief read a textfile
@@ -36,7 +36,7 @@ size_t write_file(const char *filepath, void *contents, size_t size);
  * @param out a pointer to the buffer where the contents of the text file will be written to
  * @return bytes read
  */
-size_t read_text_file(char *filepath, char **out);
+int read_text_file(char *filepath, char **out);
 
 /**
  * @brief write text to text file
@@ -46,14 +46,14 @@ size_t read_text_file(char *filepath, char **out);
  * @param size size of string
  * @return bytes written
  */
-size_t write_text_file(char *filepath, char **string, size_t size);
+int write_text_file(char *filepath, char **string, uint size);
 
 /**
  * @brief Get the size of a file
  * 
  * @param filepath path to the file
- * @return size_t in bytes
+ * @return int in bytes
  */
-size_t get_file_size(char *filepath);
+int get_file_size(char *filepath);
 
 #endif
